+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - [SwiftLint] Bump SwiftLint from 0.38.2 to 0.39.1 [#738](https://github.com/sider/runners/pull/738)
 - [FxCop] New support [#731](https://github.com/sider/runners/pull/731)
 - Aggregate log entries by `delete_unchanged_files` method [#765](https://github.com/sider/runners/pull/765)
+- [RuboCop] Use `--out=FILE` option [#769](https://github.com/sider/runners/pull/769)
 
 ## 0.20.0